+++ src/app/components/Multistaking/MultistakingForm/AmountBalanceInfo.tsx
@@ -3,14 +3,12 @@ import { useEffect } from "react";
 
 import { getNetworkConfigBTC } from "@/app/config/network/btc";
 import { usePrice } from "@/app/hooks/client/api/usePrices";
-import { useBalanceState } from "@/app/state/BalanceState";
+import { useAvailableBalance } from "@/app/hooks/services/useAvailableBalance";
 import { useMultistakingState } from "@/app/state/MultistakingState";
-import { satoshiToBtc } from "@/app/utils/btc";
 import { calculateTokenValueInCurrency } from "@/app/utils/formatCurrency";
 import { maxDecimals } from "@/app/utils/maxDecimals";
 
 export const AmountBalanceInfo = () => {
-  const { totalBtcBalance } = useBalanceState();
   const { isMaxBalanceMode, setIsMaxBalanceMode } = useMultistakingState();
 
   const btcAmount = useWatch({ name: "amount", defaultValue: "" });
@@ -19,36 +17,27 @@ export const AmountBalanceInfo = () => {
   const { coinSymbol } = getNetworkConfigBTC();
   const btcInUsd = usePrice(coinSymbol);
 
+  const { calculateAvailableBalance } = useAvailableBalance();
+  const availableBalance = calculateAvailableBalance();
+
   const btcAmountValue = parseFloat(btcAmount || "0");
   const btcAmountUsd = calculateTokenValueInCurrency(btcAmountValue, btcInUsd, {
     zeroDisplay: "$0.00",
   });
 
-  const balanceBtc = satoshiToBtc(totalBtcBalance);
-
-  const feeAmountSat = parseFloat(
-    useWatch({ name: "feeAmount", defaultValue: "0" }) || "0",
-  );
-  const feeAmountBtc = satoshiToBtc(feeAmountSat);
-
-  const availableBalanceBtc = maxDecimals(
-    Math.max(balanceBtc - feeAmountBtc, 0),
-    8,
-  );
-
   useEffect(() => {
     if (isMaxBalanceMode) {
-      setValue("amount", availableBalanceBtc.toString(), {
+      setValue("amount", availableBalance.toString(), {
         shouldValidate: true,
         shouldDirty: true,
         shouldTouch: true,
       });
     }
-  }, [isMaxBalanceMode, availableBalanceBtc, setValue]);
+  }, [isMaxBalanceMode, availableBalance, setValue]);
 
   const handleSetMaxBalance = () => {
     setIsMaxBalanceMode(true);
-    setValue("amount", availableBalanceBtc.toString(), {
+    setValue("amount", availableBalance.toString(), {
       shouldValidate: true,
       shouldDirty: true,
       shouldTouch: true,
@@ -60,7 +49,7 @@ export const AmountBalanceInfo = () => {
       <div>
         Balance:{" "}
         <u className="cursor-pointer" onClick={handleSetMaxBalance}>
-          {maxDecimals(availableBalanceBtc, 8)}
+          {maxDecimals(availableBalance, 8)}
         </u>{" "}
         {coinSymbol}
       </div>
